More fixes for resolving warnings from the Intel compiler.  These fixes resolve a subset of the remark #2259 warnings:    remark #2259: non-pointer conversion from "double" to "int" may lose significant bits
There are additional remark #2259 warnings that are yet to be addressed.

+++ met/src/basic/vx_util/interp_util.cc
@@ -314,8 +314,8 @@ double interp_bilin(const DataPlane &dp, double obs_x, double obs_y) {
    double bilin_v, dx, dy;
    double wtsw, wtse, wtnw, wtne;
 
-   x = floor(obs_x);
-   y = floor(obs_y);
+   x = nint(floor(obs_x));
+   y = nint(floor(obs_y));
 
    // Compute dx and dy
    dx = obs_x - x;
